inovice API DOne password changed issue solved

+++ controllers/admin/coupons.js
@@ -350,8 +350,10 @@ router.post('/update', async (req, res) => {
                 extention = path.extname(file.name);
                 filename = req.body.coupon_code + extention;
                 var filepath = dir + '/' + filename;
-                if(req.body.old_banner_image !== 'null'){
-                    fs.unlinkSync(dir + '/' + req.body.old_banner_image);
+                if(req.body.old_banner_image !== 'null' && req.body.old_banner_image !== null){
+                    if (fs.existsSync(dir + '/' + req.body.old_banner_image)) {
+                        fs.unlinkSync(dir + '/' + req.body.old_banner_image);
+                    }
                 }
                 file.mv(dir + '/' + filename, function (err) {
                     if (err) {
